finalizing likes
