--- public/javascripts/index.js
@@ -1,17 +1,0 @@
-document.getElementById('upvote').addEventListener('click', (e) => {
-    fetch(`/likes/1/upvote`, {
-        method: "POST",
-      })
-      .then (res => {
-          if (!res.ok) throw Error(res.statusText);
-          return res.json();
-      })
-      .then (data => {
-        //   console.log(data);
-          let score = document.getElementsByClassName('likesCount');
-          score[0].innerHTML = data.likesCount;
-      })
-      .catch (err => {
-        alert(res.message);
-      })
-});
