+++ public/stylesheets/story-id.css
@@ -277,6 +277,22 @@
   border-bottom: lightgrey solid 1px;
 }
 
+.likes{
+  display: flex;
+  justify-content: start;
+  align-items: center;
+}
+
+.likes img {
+  height: 40px;
+  width: 40px;
+}
+
+.likes p {
+  margin: 0;
+  padding: 0 0 0 7px;
+}
+
 .show{
   display:block;
 }
