+++ views/index.pug
@@ -88,4 +88,4 @@ block content
           h4 Bookmark stories for later
           p Start saving stories by clicking the bookmark icon and you’ll find them all here.
           button(id="upvote") upvote
-          h2(class="likesCount") 0
+          h2(class="likesCount")
